Moved the Account entities into a seperate folder, updated all references to this. Changed the UserAccountRepository to use the GetAccounts_byUserId stored procedure. this now returns the role id
