+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetEmployerAccount/GetEmployerAccountResponse.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using SFA.DAS.EmployerApprenticeshipsService.Domain;
+﻿using SFA.DAS.EmployerApprenticeshipsService.Domain.Entities.Account;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Application.Queries.GetEmployerAccount
 {
