+++ src/SFA.DAS.EmployerApprenticeshipsService.Application/Queries/GetUserAccounts/GetUserAccountsQueryResponse.cs
@@ -1,4 +1,5 @@
 using SFA.DAS.EmployerApprenticeshipsService.Domain;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.Entities.Account;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Application.Queries.GetUserAccounts
 {
