--- src/SFA.DAS.EmployerApprenticeshipsService.Domain/Account.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace SFA.DAS.EmployerApprenticeshipsService.Domain
-{
-    public class Account
-    {
-        public string Name { get; set; }
-        public int Id { get; set; }
-
-    }
-
-}
