+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Data/IUserAccountRepository.cs
@@ -1,4 +1,5 @@
 ﻿using System.Threading.Tasks;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.Entities.Account;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Domain.Data
 {
