+++ src/SFA.DAS.EmployerApprenticeshipsService.Domain/Entities/Account/RoleName.cs
@@ -1,0 +1,10 @@
+﻿namespace SFA.DAS.EmployerApprenticeshipsService.Domain.Entities.Account
+{
+    public enum RoleName
+    {
+        NotSet = 0,
+        Owner = 1,
+        Transactor = 1,
+        Viewer = 2
+    }
+}
