+++ src/SFA.DAS.EmployerApprenticeshipsService.Web/Controllers/VisualTestController.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Linq;
 using System.Web.Hosting;
 using System.Web.Mvc;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.Entities.Account;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Web.Controllers
 {
