+++ src/SFA.DAS.EmployerApprenticeshipsService.Web/Models/UserAccountsViewModel.cs
@@ -1,5 +1,6 @@
 using System.Collections.Generic;
 using SFA.DAS.EmployerApprenticeshipsService.Domain;
+using SFA.DAS.EmployerApprenticeshipsService.Domain.Entities.Account;
 
 namespace SFA.DAS.EmployerApprenticeshipsService.Web.Models
 {
